Replaced project and added ability to jump between planets
Continuing with the project, I accidently started working on another version of this project that was not related to this git repo so I deleted the previous version that was here and replaced it with this new version. In this new version I have added the ability for the player to jump between planets by entering their gravity area collider.