Change of plans, implemented projectile deflection
Change of plans to reduce scope of project from a survival game where you can hop between planets to a game set on one planet but you will encounter increasing enemies that fire projectiles at you. The player is able to deflect these projectiles in a direction which increases the projectiles speed, though it will still harm the player if it hits them.