Obtained projectile particle effect and stopped projectile from leaving planet
I have added a particle effect to the base enemy projectile which is one I have gotten from the Hovl Studio Magic Effects FREE pack on the unity asset store which is usable for commercial products under the Standard Unity Asset Store EULA. I have also stopped the projectile from leaving the planet when at higher speeds by increasing the gravity of the planet. I have done this with plans to scrap the ability to jump in place of a dash like ability.