Player mesh rotates towards mouse
Player now, instead of rotating in the direction of movement, rotates to look at the mouse, even while rotating to align with planet like usual. Also, the direction the projectile would go in after being deflected was sometimes off, so now instead of rotating it in the direction which I think (with higher speeds) couldn't rotate it completely, I now just set the projectiles forward direction to that of the players mesh.