Player can now die from getting it by projectile
Created the IDamageable interface which gives any class that implements it, the functionality for what to do when it takes damage. In most cases in this game, when an entity (player or enemy) is hit by a projectile, they will die. I am planning on introducing enemies with shields that allow them to take more hits and need a specifically powered projectile to break.