Enemy can now fire projectiles at player
The enemy ai has been given the ability to fire projectiles at the player which are destroyed on collision. A current issue however is that for some reason, when a projectile is spawned, the player's rotation is set to face the direction the enemy is looking (so behind the player) and if the player moves during this time, the rotation goes haywire, causing the player to spin around randomly.