Created the scoring system and added TMP assets
I have created the working scoring system for this game. How it works is it consists of two parts, the "temporary" score which the player is able to lose and the "banked" score which can only be gained. When the player deflects a projectile, they gain temporary score and if they defeat an enemy, their temporary score is banked. However, if the player goes too long without deflecting any projectiles or they are hit by a projectile, they lose their temporary score and it goes back to 0.