Added player death animation, enemy idle animation and fixed projectile redirection
From mixamo.com, I have obtained a death animation for the player where they are sent flying backwards a bit. This was imported into unity and implemented so it plays on player death. I have also added an enemy idle animation for when the player dies and before the enemy hits the ground after they spawn. This is to mainly stop them from continuously firing even after they have killed the player. The deflection direction for the players deflection was also off as it was set to the player mesh's forward vector however that was causing the projectile direction to be off. Instead I just use the player mesh-to-mouse direction as the new facing direction for the projectile.