Fixed directional movement
For the majority of this projects development, the locomotion animation played was solely dependant on what directional key you were pressing such as holding W always showing the move forward animation even if the player character was moving backwards. This has now been changed so the animation played is dependant on the angle between the players forward vector and their movement vector.