Projectiles now award points based off stage and stronger enemies spawned over time
When the player deflects a projectile, the higher the stage of it, the more temporary points are awarded to the player. This can incentivise the player to go for a higher rally as it will award them more points if they can defeat an enemy afterwards. Over time, as the internal wave number increases, enemies with stronger shield will spawn instead of the lower ones. This means players will have to build up a stronger projectile to defeat them. This adds ramping difficulty to the game as it would get boring quickly with just the base enemies.