Added the ability to pause the game and other minor adjustments
I have now created a basic pause menu which can be accessed by the player by them pressing the ESC button. From the pause menu they can either resume the game or they can return to the main menu. I still need to find a suitable font for all of the UI. I also created a dissolve shader for when enemies die and I think I have finally fixed the bug of deflected projectiles not going in the right direction while the player is at certain parts of the planet (at certain angles).