Added failure menu, background music, fade ins/outs for music and scenes
Added a menu for when the player eventually dies which shows their score at the end and a button that allows them to either reload the level and start again or quit back to the main menu. I have also obtained some nice fitting background music from PlayOnLoop.com called "Last Samurai" which plays in the main menu and during actual gameplay. For somewhat better feel, I have made the music and gameplay scene fade in and out when necessary.